documents download from local and hide upload from docuware tab

+++ resources/views/projects/partials/local-documents.blade.php
@@ -71,12 +71,12 @@
 	             		<span class="a-pencil-2" style="color: rgb(0, 193, 247);"></span>
 	            </a>
 	            &nbsp;|&nbsp;
-			    	 {{-- @can('access_admin') --}}
+			    	 @can('access_admin')
 			    		<a class="uk-link-muted " onclick="deleteFile({{ $document->id }});" uk-tooltip="Delete this file">
 			    			<span class="a-trash-4" style="color: #da328a;"></span>
 			    		</a>
 			    		&nbsp;|&nbsp;
-			    	 {{-- @endcan --}}
+			    	 @endcan
 			    		<a href="download-local-document/{{ $document->id }}" target="_blank"  uk-tooltip="Download file." download>
 			    			<span class="a-lower"></span>
 			    		</a>
